Basic bond to topology conversion
