+++ common/client-libs/validator-client-rest/Cargo.toml
@@ -12,4 +12,3 @@ mixnet-contract = { path = "../../../common/mixnet-contract" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 reqwest = { version = "0.11", features = ["json"] }
-
