+++ common/client-libs/validator-client/src/models/gateway.rs
@@ -18,25 +18,7 @@ use serde::{Deserialize, Serialize};
 use std::convert::TryInto;
 use std::io;
 use std::net::{SocketAddr, ToSocketAddrs};
-
-#[derive(Debug)]
-pub enum ConversionError {
-    InvalidIdentityKeyError(identity::KeyRecoveryError),
-    InvalidSphinxKeyError(encryption::KeyRecoveryError),
-    InvalidAddress(io::Error),
-}
-
-impl From<encryption::KeyRecoveryError> for ConversionError {
-    fn from(err: encryption::KeyRecoveryError) -> Self {
-        ConversionError::InvalidSphinxKeyError(err)
-    }
-}
-
-impl From<identity::KeyRecoveryError> for ConversionError {
-    fn from(err: identity::KeyRecoveryError) -> Self {
-        ConversionError::InvalidIdentityKeyError(err)
-    }
-}
+use topology::gateway::GatewayConversionError;
 
 // used for gateways to register themselves
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
@@ -102,15 +84,15 @@ impl RegisteredGateway {
         &self.gateway_info.node_info.version
     }
 
-    fn resolve_hostname(&self) -> Result<SocketAddr, ConversionError> {
+    fn resolve_hostname(&self) -> Result<SocketAddr, GatewayConversionError> {
         self.gateway_info
             .node_info
             .mix_host
             .to_socket_addrs()
-            .map_err(ConversionError::InvalidAddress)?
+            .map_err(GatewayConversionError::InvalidAddress)?
             .next()
             .ok_or_else(|| {
-                ConversionError::InvalidAddress(io::Error::new(
+                GatewayConversionError::InvalidAddress(io::Error::new(
                     io::ErrorKind::Other,
                     "no valid socket address",
                 ))
@@ -119,7 +101,7 @@ impl RegisteredGateway {
 }
 
 impl TryInto<topology::gateway::Node> for RegisteredGateway {
-    type Error = ConversionError;
+    type Error = GatewayConversionError;
 
     fn try_into(self) -> Result<topology::gateway::Node, Self::Error> {
         Ok(topology::gateway::Node {
@@ -138,7 +120,7 @@ impl TryInto<topology::gateway::Node> for RegisteredGateway {
 }
 
 impl<'a> TryInto<topology::gateway::Node> for &'a RegisteredGateway {
-    type Error = ConversionError;
+    type Error = GatewayConversionError;
 
     fn try_into(self) -> Result<topology::gateway::Node, Self::Error> {
         Ok(topology::gateway::Node {
