+++ common/mixnet-contract/Cargo.toml
@@ -10,3 +10,4 @@ edition = "2018"
 cosmwasm-std = { version = "0.13.2" }
 serde = { version = "1.0", features = ["derive"] }
 schemars = "0.7"
+topology = { path = "../topology"}
